backend/requirements.txt
fastapi  
uvicorn  
pydantic  
motor  
python-jose  
passlib[bcrypt]